requirements.txt
django==5.0.6
django-cors-headers==4.3.1
djangorestframework==3.15.1
django-ses==4.1.0
Pillow==10.3.0
djoser==2.2.2
argon2-cffi==23.1.0
psycopg2==2.9.9
requests==2.32.2
whitenoise==6.6.0
python-decouple==3.8